modified:   client/src/pages/Create.jsx     - The form structure has been adjusted.     - Added alert for form validation.
modified:   client/src/pages/HomeContent.jsx
    - Next and Prev buttons were removed.

modified:   client/src/pages/SinglePost.jsx
    - Confirm message added before post is deleted

+++ client/src/pages/SinglePost.jsx
@@ -30,6 +30,10 @@ export default function SinglePost() {
     }
 
     const deletePostId = async () => {
+        const confirm = window.confirm("Bu bloğu silmek istediğinizden emin misiniz?");
+
+        if (!confirm) return;
+
         try {
             await dispatch(deletePostAsyncThunk(match.params.id));
             navigate("/");
@@ -44,9 +48,8 @@ export default function SinglePost() {
             <div className="container px-4 px-lg-5">
                 {
                     token && (
-                        <div className="d-flex justify-content-end mb-4">
-                            <button className="btn btn-outline-primary me-3">Update Post</button>
-                            <button onClick={() => deletePostId()} className="btn btn-outline-danger">Delete Post </button>
+                        <div className="d-flex justify-content-start mb-4">
+                            <button onClick={() => deletePostId()} className="btn btn-danger" style={{ width: "10%"}}>Sil</button>
                         </div>
                     )
                 }
